Fix type conversion in `CustomTotemRessurectEffect` for compatibility

+++ src/main/kotlin/org/ReDiego0/animationPlayer/utils/effects/CustomTotemRessurectEffect.kt
@@ -32,7 +32,7 @@ object CustomTotemRessurectEffect {
 
         packet.integers.write(0, 0)
         packet.integers.write(1, 0)
-        packet.shorts.write(0, 45.toShort())
+        packet.shorts.write(0, 45)
         packet.itemModifier.write(0, itemStack)
 
         try {
